Refactor value lowerer
